requirements.txt
streamlit
python-dotenv
openai>=1.0.0
pydub
reportlab
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
